requirements.txt
git+https://github.com/Farama-Foundation/Metaworld.git@master#egg=metaworld
gymnasium
gymnasium[other]
hydra-core
wandb
jax
jax[cuda12]
flax
distrax
optax
scikit-learn
tensorflow-probability
sentence-transformers